backend/requirements.txt
# Web Framework & Server
flask>=3.0.0                          # Lightweight web framework for REST API endpoints
gunicorn>=21.0.0                      # Production WSGI HTTP server for deploying Flask app

# Configuration & Environment
python-dotenv>=1.0.0                  # Load environment variables from .env file
cryptography>=40.0.0                  # Encryption/hashing for API keys and sensitive data

# Data Validation & Serialization
pydantic>=2.0.0                       # Type-safe request/response validation using Python type hints

# AI Orchestration & LangGraph
langchain>=0.2.0                      # LLM orchestration, tool calling, and chain management
langgraph>=0.1.0                      # State machine for multi-agent workflow (core to our 5-agent pipeline)
langchain-groq>=0.1.0                 # Groq LLM integration for ultra-fast agents
langchain-community>=0.0.10           # Community LLM integrations

# Databases & Caching
pymongo>=4.0.0                        # MongoDB driver for storing chat history and graph state snapshots
diskcache>=5.0.0                      # Disk-based caching for query results and LLM responses

# HTTP & Utilities
requests>=2.28.0                      # HTTP library for MCP tools (web search, API calls)
